% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Takes the euclidian algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{shall be a numeric value and a scalar}

\item{b}{shall be a numeric value and a scalar}
}
\value{
Returns a numeric scalar that is the greatest common divisor of a and b.
}
\description{
An algorithm that computes the greatest common divisor (GDC) of two numeric numbers.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
\references{
The info about the algorithm can be found at this link  
\href{https://en.wikipedia.org/wiki/Euclidean_algorithm}{wikepedia.}
}
