% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab3AlbinEric.R
\docType{package}
\name{Lab3AlbinEric}
\alias{Lab3AlbinEric}
\alias{Lab3AlbinEric-package}
\title{Lab3AlbinEric: A package for the algorithms dijkstra and euclidean}
\description{
The Lab3AlbinEric package provides three categories of important functions:
}
\section{Lab3AlbinEric functions}{

dijkstra
euclidean
}

\section{Lab3AlbinEric data}{

wiki_graph
}

\author{
Albin Vasterlund and Eric Herwin
}
