% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_x.R
\name{power_x}
\alias{power_x}
\title{Gives u the power of x}
\usage{
power_x(x)
}
\arguments{
\item{x}{shall be a numeric value}
}
\description{
This funcktion will give u the power of x
}
\examples{
\dontrun{
# Dont run this
}
power_x(x=10)
}
