% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Length of paths betwwen differend nodes.}
\format{A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{number of the first node}
  \item{v2}{number of the second node that the first node goes to}
  \item{w}{the distance between v1 and v2}
}}
\source{
\url{https://en.wikipedia.org/wiki/Graph}
}
\usage{
wiki_graph
}
\description{
A dataset containing 6 different nodes and the distance betwwen one another.
}
\keyword{datasets}
